Rename resolver package into environment
