+++ core/src/main/java/com/nikodoko/javaimports/environment/maven/MavenDependencyVersion.java
@@ -1,4 +1,4 @@
-package com.nikodoko.javaimports.resolver.maven;
+package com.nikodoko.javaimports.environment.maven;
 
 import com.google.common.base.MoreObjects;
 import java.util.Objects;
