+++ core/src/main/java/com/nikodoko/javaimports/fixer/internal/Loader.java
@@ -5,8 +5,8 @@ import com.nikodoko.javaimports.parser.ClassHierarchies;
 import com.nikodoko.javaimports.parser.ClassHierarchy;
 import com.nikodoko.javaimports.parser.Import;
 import com.nikodoko.javaimports.parser.ParsedFile;
-import com.nikodoko.javaimports.resolver.Resolver;
-import com.nikodoko.javaimports.resolver.Resolvers;
+import com.nikodoko.javaimports.environment.Resolver;
+import com.nikodoko.javaimports.environment.Resolvers;
 import com.nikodoko.javaimports.stdlib.StdlibProvider;
 import com.nikodoko.javaimports.stdlib.StdlibProviders;
 import java.util.HashMap;
