+++ core/src/test/java/com/nikodoko/javaimports/environment/PackageDistanceTest.java
@@ -1,4 +1,4 @@
-package com.nikodoko.javaimports.resolver;
+package com.nikodoko.javaimports.environment;
 
 import static com.google.common.truth.Truth.assertThat;
