+++ core/src/test/java/com/nikodoko/javaimports/environment/maven/MavenDependencyVersionTest.java
@@ -1,4 +1,4 @@
-package com.nikodoko.javaimports.resolver.maven;
+package com.nikodoko.javaimports.environment.maven;
 
 import static com.google.common.truth.Truth8.assertThat;
